requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
paho-mqtt==1.6.1
httpx==0.25.2
python-multipart==0.0.6
